pyproject.toml
[project]
name = "email-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "mail-parser>=4.1.4",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "pdfminer-six>=20251107",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "python-docx>=1.2.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "rank-bm25>=0.2.2",
    "scipy>=1.16.3",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "tinydb>=4.8.2",
    "uvicorn[standard]>=0.38.0",
]
